Up code

+++ src/java/helpers/encrypt.java
@@ -17,8 +17,9 @@ public class encrypt {
             throw new RuntimeException(e);
         }               
     }
+    
     public static void main(String[] args) throws Exception {
-        double b = 1000000000;
+        int b = 1000000000;
         System.out.println("Xuất: " + b);
         System.out.println("Kết quả đã mã hóa: " + hashmd5("demo@gmail.com", "demo"));
     }
